address PR feedback for #44
Remove modifications to existing ShowHelp methods and instead add
overloads that take a bool argument for inheritances. In order to ensure
proper inheritance all the way down the tree of subcommands, the Command
method now manually sets the OptionHelp property of the created sub command
after configuration.
Make changes in test to assert the expected help text is actually written to the console.