Adds a test for validation of subcommand models
This commit demonstates a failing unit test for command validation in subcommands.

Currently the theories on lines 99, 100, and 101 in ValidateMethodConventionTest.cs fail when they should pass (if I understand the intention correctly). The tests on lines 102, 103, and 104 pass and should pass.

The root cause seems to be that an initialized model, **but not a model that has had its values bound to args**, is used for validation in subcommands.

For main commands (not subcomands) the model used for validation is initialized and has had its values bound - as expected.