Handle nested HelpOptions with inheritance
When the HelpOption with  is set with inherited = true and nested subcommands set the HelpOption too then the HelpOption nearest to the selected command is used.

Extracted the search for options by ShortName , SymbolName, LongName into a new method called FindOption.

Added Unit Tests for each branch that calls the FindOption method.