Enable overridding console width of the DefaultHelpTextGenerator
Used e.g. for testing.

Also made the default constructor public. This allows the DefaultHelpTextGenerator to be instantiated on a per-use basis in order to keep unit tests a isolated from each other as possible.
Without that change, all instances would be the same singleton instance - i.e. a test can affect another test by fiddling with the overridden width.